# Use strings.Builder for querybuilder build method