It's a minor improvement to `querybuilder` performance, given this benchmark:

```go
func BenchmarkBuilder(b *testing.B) {
	for i := 0; i < b.N; i++ {
		var contents string
		root := Query().
			Select("foo").
			Select("bar").Arg("hello", "world").
			Select("field").Arg("test", "test").Bind(&contents)
		root.build(context.Background())
		_, _ = Query().
			Select("a").Arg("arg", "one").
			Select("b").Arg("arg", "two").
			build(context.Background())
	}
}
```

`benchcmp` results:
```
% benchcmp old new 
benchcmp is deprecated in favor of benchstat: https://pkg.go.dev/golang.org/x/perf/cmd/benchstat
benchmark                      old ns/op     new ns/op     delta
BenchmarkBuilder-8     11363         10839         -4.61%

benchmark                      old allocs     new allocs     delta
BenchmarkBuilder-8     88             76             -13.64%

benchmark                      old bytes     new bytes     delta
BenchmarkBuilder-8     3232          2849          -11.85%
```